index.jsの記述をapp.jsに移動
